+++ frontend/client/public/manifest.json
