+++ frontend/client/src/App.js
@@ -1,0 +1,8 @@
+import './App.css';
+
+function App() {
+  return (
+    <div className="app"></div>);
+}
+
+export default App;
