--- frontend/error404.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-function Error404() {
-  return (
-    <div>
-      <h1>404 Error</h1>
-      <p>Sorry, the page you are looking for does not exist.</p>
-      <img src="https://pbs.twimg.com/media/EAmr-PAWsAEoiWR.jpg" alt="SadCat" />
-    </div>
-  );
-}
-
-export default Error404;
